Abstract describe block
